Refactor to move all party use cases into a party folder
